pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "CADET-Python-Simulator"
dynamic = ["version"]
authors = [
    { name = "Johannes Schmölder", email = "j.schmoelder@fz-juelich.de" },
    { name = "Daniel Klauß", email = "d.klauss@fz-juelich.de" },
    { name = "Antonia Berger", email = "a.berger@fz-juelich.de" },
    { name = "Eric von Lieres", email = "e.von.lieres@fz-juelich.de" }
]
description = "A Python toolbox for simulating systems of unit operations."
readme = "README.md"
requires-python = ">3.10"
keywords = ["research data management"]
license = { text = "GPLv3" }
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Intended Audience :: Science/Research",
]
dependencies = [
    "addict",
    "numpy",
    "scipy",
    "matplotlib",
    "scikits.odes",
    "cadet-process",
]

[project.urls]
homepage = "https://jugit.fz-juelich.de/IBG-1/ModSim/cadet/CADET-Python-Simulator"
"Bug Tracker" = "https://jugit.fz-juelich.de/IBG-1/ModSim/cadet/CADET-Python-Simulator/Issues"


[project.optional-dependencies]
testing = [
    "setuptools",
    "pytest",
    "certifi", # tries to prevent certificate problems on windows
    "pre-commit", # system tests run pre-commit
    "flake8", # system tests run flake8
    "build",
]
docs = [
    "sphinx>=5.3.0",
    "sphinxcontrib-bibtex>=2.5.0",
    "sphinx_book_theme>=1.0.0",
    "sphinx_copybutton>=0.5.1",
    "sphinx-sitemap>=2.5.0",
    "numpydoc>=1.5.0",
    "myst-nb>=0.17.1",
]


[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "server_api: marks tests as using the GitLab/GitHub API"
]

[tool.setuptools.dynamic]
version = { attr = "CADETPythonSimulator.__version__" }

[tool.ruff]
# Same as Black.
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = ["E", "F", "W", "D"]
ignore = ["F401", "D212", "D100"]
